Power Prerequisite Endpoint migrated to permissions
